refactor!: move create_(anti)particle to data
